# modules/openapi/src/commonMain/kotlin/com/sphereon/oid/fed/openapi/admin-server.yaml
openapi: 3.0.3
info:
  title: OpenID Federation API
  description: This API allows the interaction with federation endpoints to manage and verify entities within a federation.
  contact:
    name: Sphereon
    email: info@sphereon.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0-d40
servers:
  - description: SwaggerHub API Auto Mocking
    url: 'https://virtserver.swaggerhub.com/SphereonInt/OpenIDFederationAPI/1.0.0-d40'
  - url: http://localhost:8081
    description: Admin Server
paths:
...
  /accounts:
    get:
      summary: List all accounts
      description: Retrieve a list of all accounts.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      responses:
        '200':
          description: Accounts retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Register a new tenant account
      description: Endpoint for an admin to create a new account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccount'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'Conflict (e.g., account already exists)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete current account
      description: Delete the account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      responses:
        '200':
          description: Account deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/keys':
    post:
      summary: Create a new key
      description: Create a new key for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      responses:
        '201':
          description: Key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jwk'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List all keys
      description: Retrieve all keys associated with the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      responses:
        '200':
          description: Keys retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Jwk'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/keys/{keyId}':
    delete:
      summary: Revoke a key
      description: Revoke a key by its ID for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: keyId
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the key to be revoked.
        - name: reason
          in: query
          required: false
          schema:
            type: string
          description: The reason for revoking the key.
      responses:
        '200':
          description: Key revoked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jwk'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Key or account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/metadata':
    get:
      summary: Get entity configuration metadata
      description: Retrieve metadata entries for a specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      responses:
        '200':
          description: Successfully retrieved metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Metadata'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create entity configuration metadata
      description: Add a new metadata entry for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      requestBody:
        description: Metadata creation data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMetadata'
      responses:
        '201':
          description: Metadata created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/metadata/{id}':
    delete:
      summary: Delete account metadata
      description: Deletes a specific metadata entry for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the metadata entry to be deleted.
      responses:
        '200':
          description: Metadata deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Metadata not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/authority-hints':
    get:
      summary: Get authority hints
      description: Retrieve all authority hints for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      responses:
        '200':
          description: Successfully retrieved authority hints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthorityHint'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create an authority hint
      description: Add a new authority hint for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      requestBody:
        description: Authority hint data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthorityHint'
      responses:
        '201':
          description: Authority hint created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorityHint'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/authority-hints/{id}':
    delete:
      summary: Delete an authority hint
      description: Remove an authority hint by its ID for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the authority hint to be deleted.
      responses:
        '200':
          description: Authority hint deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorityHint'
        '404':
          description: Authority hint or account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/entity-statement':
    get:
      summary: Get entity configuration statement object
      description: Retrieve the entity configuration statement for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      responses:
        '200':
          description: Successfully retrieved entity configuration statement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityConfigurationStatement'
        '404':
          description: Account or entity statement not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Sign and publish entity configuration statement
      description: Sign and publish the entity configuration statement for the specified account. If `dry-run` is `true`, outputs the signed JWT without publishing it.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      requestBody:
        description: Optional request body for a dry-run
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishEntityStatement'
      responses:
        '200':
          description: Entity configuration statement published or validated successfully
          content:
            application/json:
              schema:
                type: string
                description: A message indicating the result of the operation.
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  '/subordinates/{subordinateId}/metadata/{id}':
    delete:
      summary: Delete subordinate metadata
      description: Delete metadata associated with the specified subordinate.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: subordinateId
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the subordinate.
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: The ID of the metadata to be deleted.
      responses:
        '200':
          description: Subordinate metadata deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubordinateMetadata'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    AccountUsername:
      name: X-Account-Username
      in: header
      description: Optional account username to operate on behalf of
      required: false
      schema:
        type: string
...
  schemas:
    Account:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier for the account.
          example: 12345
        username:
          type: string
          description: The username of the account.
          example: acmecorp
        identifier:
          type: string
          description: The identifier of the account.
          format: uri
          example: "https://www.example.com/oidf"
          pattern: "^https?:\\/\\/.*$"
      required:
        - id
        - username
    AuthorityHint:
      type: object
      properties:
        id:
          type: integer
          format: int32
        identifier:
          type: string
        accountId:
          type: integer
          format: int32
      required:
        - id
        - identifier
        - accountId
        - admin
    BaseJwk:
      type: object
      x-tags:
        - federation
      required:
        - kty
        - kid
      properties:
        kty:
          type: string
          description: 'The key type (e.g., EC).'
          example: EC
        crv:
          type: string
          description: The elliptic curve used (only for EC keys).
          example: P-256
          nullable: true
        kid:
          type: string
          description: The key ID.
          example: 12345
          nullable: true
        x:
          type: string
          description: The X coordinate for EC keys (optional).
          example: o-7zraXKDaoBte2PsuTXo-MSLzsyWdAElNptGgI4aH8
          nullable: true
        y:
          type: string
          description: The Y coordinate for EC keys (optional).
          example: Xr_wCzJ1XnsgAIV5qHruzSwaNnwy87UjmevVklTpIv8
          nullable: true
        n:
          type: string
          description: The modulus for RSA keys.
          example: modulus_value
          nullable: true
        e:
          type: string
          description: The exponent for RSA keys.
          example: AQAB
          nullable: true
        alg:
          type: string
          description: The algorithm associated with the key.
          example: ES256
          nullable: true
        use:
          type: string
          description: 'The intended use of the key (e.g., sig, enc).'
          example: sig
          nullable: true
        x5u:
          type: string
          format: uri
          description: A URL that points to an X.509 public key certificate or certificate chain.
          example: 'https://example.com/cert.pem'
          nullable: true
        x5c:
          type: array
          description: The X.509 certificate chain.
          nullable: true
          items:
            type: string
            description: A base64-encoded string representing an X.509 certificate.
            example: MIICoTCCAYkCAQ...
        x5t:
          type: string
          description: The SHA-1 thumbprint of the X.509 certificate.
          example: dGhpcyBpcyBqdXN0IGEgdGh1bWJwcmludA
          nullable: true
        x5tS256:
          type: string
          description: The SHA-256 thumbprint of the X.509 certificate.
          example: sM4KhEI1Y2Sb6-EVr6tJabmJuoP-ZE...
          nullable: true
    BaseStatement:
      type: object
      required:
        - iss
        - sub
        - iat
        - exp
        - jwks
      properties:
        iss:
          type: string
          description: The Entity Identifier of the issuer of the Entity Statement.
        sub:
          type: string
          description: The Entity Identifier of the subject.
        exp:
          type: integer
          description: Expiration time after which the statement MUST NOT be accepted for processing.
        iat:
          type: integer
          format: date-time
          description: The time the statement was issued.
        jwks:
          type: object
          x-tags:
            - federation
          properties:
            keys:
              type: array
              items:
                $ref: '#/components/schemas/Jwk'
        metadata:
          additionalProperties: true
        crit:
          type: array
          items:
            type: string
    CreateAccount:
      type: object
      properties:
        username:
          type: string
          description: The username of the account.
          example: acmeco
        identifier:
          type: string
          description: The identifier of the tenant account.
          example: https://www.example.com/oidf
      required:
        - username
    CreateAuthorityHint:
      type: object
      properties:
        identifier:
          type: string
          description: The identifier of the authority hint.
      required:
        - identifier
...
    CreateMetadata:
      type: object
      properties:
        key:
          type: string
          description: The metadata key.
          example: openid_relying_party
        metadata:
          additionalProperties: true
          description: The metadata object.
      required:
        - key
        - metadata
...
    EntityConfigurationStatement:
      allOf:
        - $ref: '#/components/schemas/BaseStatement'
        - type: object
          properties:
            authority_hints:
              type: array
              items:
                type: string
            trust_marks:
              type: array
              description: 'An array of JSON objects, each representing a Trust Mark.'
              items:
                $ref: '#/components/schemas/TrustMark'
            trust_mark_issuers:
              $ref: '#/components/schemas/TrustMarkIssuers'
            trust_mark_owners:
              $ref: '#/components/schemas/TrustMarkOwners'
    ErrorResponse:
      type: object
      x-tags:
        - federation
      required:
        - error
        - message
      properties:
        status:
          type: integer
          x-stoplight:
            id: joax1cgth4uzd
        error:
          type: string
        message:
          type: string
          description: A human-readable short text describing the error.
        timestamp:
          type: integer
          x-stoplight:
            id: qtn6mqbzpjctv
...
    Jwk:
      allOf:
        - $ref: '#/components/schemas/BaseJwk'
        - type: object
          x-tags:
            - federation
          properties:
            id:
              type: integer
              description: The unique identifier for the JWK record.
              example: 1
            uuid:
              type: string
              format: uuid
              description: The universally unique identifier for the JWK record.
              example: 123e4567-e89b-12d3-a456-426614174000
            account_id:
              type: integer
              description: The ID of the account associated with this JWK.
              example: 100
            revoked_at:
              type: string
              format: date-time
              description: 'The timestamp when the JWK was revoked, if applicable.'
              example: '2024-09-01T12:34:56Z'
              nullable: true
            revoked_reason:
              type: string
              description: 'The reason for revoking the JWK, if applicable.'
              example: Key compromise
              nullable: true
            created_at:
              type: string
              format: date-time
              description: The timestamp when the JWK was created.
              example: '2024-08-06T12:34:56Z'
              nullable: true
...
    Metadata:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier for the Metadata record.
          example: 1
        account_id:
          type: integer
          description: The ID of the account associated with this Metadata.
          example: 1
        key:
          type: string
          description: The key of the metadata.
          example: openid_relying_party
        metadata:
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          description: The timestamp when the Metadata was created.
          example: '2024-08-06T12:34:56Z'
          nullable: false
        deleted_at:
          type: string
          format: date-time
          description: The timestamp when the Metadata was deleted.
          example: '2024-08-06T12:34:56Z'
          nullable: true
    PublishEntityStatement:
      type: object
      x-tags:
        - federation
      properties:
        dry-run:
          type: boolean
          description: 'If true, the request will be validated but not persisted.'
          example: false
...
    TrustMark:
      type: object
      x-tags:
        - federation
      properties:
        id:
          type: string
          description: The Trust Mark identifier. It MUST be the same value as the id claim contained in the Trust Mark JWT.
          example: example-trust-mark-id
        trust_mark:
          type: string
          description: A signed JSON Web Token that represents a Trust Mark.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
      required:
        - id
        - trust_mark
    TrustMarkIssuers:
      type: object
      x-tags:
        - federation
      description: A mapping of trust mark identifiers to their associated issuers.
      additionalProperties:
        type: array
        description: A list of issuers for the trust mark.
        items:
          type: string
          format: uri
          description: The URI of an issuer for the trust mark.
      example:
        'https://openid.net/certification/op':
          - 'https://example-issuer1.com'
          - 'https://example-issuer2.com'
        'https://refeds.org/wp-content/uploads/2016/01/Sirtfi-1.0.pdf':
          - 'https://swamid.se'
    TrustMarkOwners:
      type: object
      x-tags:
        - federation
      additionalProperties:
        $ref: '#/components/schemas/TrustMarkOwner'
      example:
        'https://refeds.org/wp-content/uploads/2016/01/Sirtfi-1.0.pdf':
          sub: 'https://refeds.org/sirtfi'
          jwks:
            keys:
              - alg: RS256
                e: AQAB
                kid: key1
                kty: RSA
                'n': pnXBOusEANuug6ewezb9J_...
                use: sig
    TrustMarkOwner:
      type: object
      x-tags:
        - federation
      properties:
        sub:
          type: string
          description: Identifier of the Trust Mark owner
        jwks:
          type: array
          items:
            $ref: '#/components/schemas/BaseJwk'
        additionalProperties:
          type: string
...
    SubordinateMetadata:
      type: object
      x-tags:
        - federation
      properties:
        id:
          type: integer
          description: The unique identifier for the Subordinate Metadata record.
          example: 1
        account_id:
          type: integer
          description: The ID of the account associated with this Metadata.
          example: 1
        subordinate_id:
          type: integer
          description: The ID of the subordinate associated with this Metadata.
          example: 1
        key:
          type: string
          description: The key of the metadata.
          example: openid_relying_party
        metadata:
          additionalProperties: true
        created_at:
          type: string
          format: date-time
          description: The timestamp when the Metadata was created.
          example: '2024-08-06T12:34:56Z'
          nullable: false
        deleted_at:
          type: string
          format: date-time
          description: The timestamp when the Metadata was deleted.
          example: '2024-08-06T12:34:56Z'
          nullable: true
...